refactor(Piece): 추상메서드를 일반 메서드로 변경하여 하위 메서드들이 오버라이드 할 필요 없게 변경
