+++ src/main/java/chess/domain/piece/King.java
@@ -26,16 +26,6 @@ public class King extends Piece {
     }
 
     @Override
-    public boolean isBlank() {
-        return false;
-    }
-
-    @Override
-    public boolean isPawn() {
-        return false;
-    }
-
-    @Override
     public boolean isKing() {
         return true;
     }
@@ -43,15 +33,5 @@ public class King extends Piece {
     @Override
     public double score() {
         return KING_SCORE;
-    }
-
-    @Override
-    public boolean diagonal(Position from, Position to) {
-        throw new InvalidMethodCallException();
-    }
-
-    @Override
-    public boolean forward(Position from, Position to) {
-        throw new InvalidMethodCallException();
     }
 }
